ROLLBACK: Revert to original App.tsx - ic-use-internet-identity broken
## Critical Issue
ic-use-internet-identity library causes runtime error:
TypeError: e.getTimeDiffMsecs is not a function

Login button non-functional, page crashes on load.

## Rollback Applied
✅ Restored original App.tsx (manual actor management)
✅ Restored original useICPI.ts (actor as props)
✅ Deployed working version to mainnet
✅ Site functional again: https://qhlmp-5aaaa-aaaam-qd4jq-cai.icp0.io

## Files Preserved
- src/App-broken.tsx (refactored version with error)
- src/hooks/useICPI-broken.ts (refactored version)
- src/hooks/actors/* (still in codebase for future)

## Root Cause Investigation Needed
The ic-use-internet-identity library used by alex_frontend doesn't work
in our build environment. Possible causes:
1. Version incompatibility
2. Missing peer dependencies
3. Vite build configuration issue
4. Need different initialization pattern

## Alternative Approaches
1. Use ic-use-actor WITHOUT ic-use-internet-identity (manual auth)
2. Match exact alex_frontend dependency versions
3. Use @dfinity/auth-client directly with better patterns
4. Hybrid: Keep manual auth, use ic-use-actor for backend only

## Status
🔴 ic-use-actor refactor BLOCKED - needs investigation
✅ Original version working on mainnet
⏳ Still have 20-second delays and missing data (original issues)
