+++ src/icpi_frontend/src/hooks/useICPI.ts
@@ -1,8 +1,6 @@
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
-import { HttpAgent, Actor } from '@dfinity/agent'
+import { Actor, HttpAgent } from '@dfinity/agent'
 import { Principal } from '@dfinity/principal'
-import { useInternetIdentity } from 'ic-use-internet-identity'
-import { useICPIBackend } from './actors'
 import { ICPI_BACKEND_CANISTER_ID, ICPI_TOKEN_CANISTER_ID, CKUSDT_CANISTER_ID } from '../constants/canisters'
 
 // Types
@@ -30,28 +28,6 @@ export const QUERY_KEYS = {
   USER_WALLET_BALANCES: 'userWalletBalances',
   TOTAL_SUPPLY: 'totalSupply',
 } as const
-
-// Helper to create authenticated agent
-function useAuthenticatedAgent(): HttpAgent | null {
-  const { identity } = useInternetIdentity()
-
-  if (!identity) return null
-
-  const isLocal = window.location.hostname === 'localhost' || window.location.hostname === '127.0.0.1'
-  const host = isLocal ? 'http://localhost:4943' : 'https://icp-api.io'
-
-  const agent = new HttpAgent({
-    identity,
-    host,
-    ingressExpiryMs: 5 * 60 * 1000,
-  })
-
-  if (isLocal) {
-    agent.fetchRootKey().catch(console.error)
-  }
-
-  return agent
-}
 
 // ICRC1/ICRC2 IDL for balance queries, transfers, and approvals
 const ICRC1_IDL = ({ IDL }: any) => {
@@ -123,18 +99,17 @@ const ICRC1_IDL = ({ IDL }: any) => {
   })
 }
 
-// ===== QUERY HOOKS =====
-
-export const useIndexState = () => {
-  const { actor } = useICPIBackend()
-
+// Custom hooks for ICPI data
+export const useIndexState = (actor: Actor | null) => {
   return useQuery({
     queryKey: [QUERY_KEYS.INDEX_STATE],
     queryFn: async () => {
       if (!actor) throw new Error('Actor not initialized')
 
+      // Try cached version first (fast query)
       const result = await actor.get_index_state_cached()
 
+      // Handle Result type - unwrap Ok/Err variant
       if ('Ok' in result) {
         return result.Ok
       } else if ('Err' in result) {
@@ -144,21 +119,21 @@ export const useIndexState = () => {
       throw new Error('Unexpected result format')
     },
     enabled: !!actor,
-    refetchInterval: 60_000,
-    staleTime: 30_000,
+    refetchInterval: 60_000, // Refetch every 60 seconds (cache refreshes every 5 min on backend)
+    staleTime: 30_000, // Consider stale after 30 seconds
     retry: 3,
     retryDelay: (attemptIndex) => Math.min(1000 * 2 ** attemptIndex, 30000),
+    onError: (error: any) => console.error('useIndexState error:', error),
   })
 }
 
-export const useRebalancerStatus = () => {
-  const { actor } = useICPIBackend()
-
+export const useRebalancerStatus = (actor: Actor | null) => {
   return useQuery({
     queryKey: [QUERY_KEYS.REBALANCER_STATUS],
     queryFn: async () => {
       if (!actor) throw new Error('Actor not initialized')
 
+      // Add timeout wrapper to prevent hanging
       const timeoutPromise = new Promise((_, reject) => {
         setTimeout(() => {
           console.warn('⚠️ get_rebalancer_status timed out after 10s')
@@ -174,22 +149,22 @@ export const useRebalancerStatus = () => {
       return result
     },
     enabled: !!actor,
-    refetchInterval: 60_000,
+    refetchInterval: 60_000, // Refetch every 1 minute
     staleTime: 30_000,
-    retry: 0,
-    retryDelay: 1000,
+    retry: 0, // Don't retry on timeout
+    retryDelay: 1000, // Wait 1s between retries
+    onError: (error: any) => console.error('useRebalancerStatus error:', error),
   })
 }
 
-export const useTVLData = () => {
-  const { actor } = useICPIBackend()
-
+export const useTVLData = (actor: Actor | null) => {
   return useQuery({
     queryKey: [QUERY_KEYS.TVL_DATA],
     queryFn: async () => {
       if (!actor) throw new Error('Actor not initialized')
       const result = await actor.get_tvl_summary()
 
+      // Handle Result type - unwrap Ok/Err variant
       if ('Ok' in result) {
         return result.Ok
       } else if ('Err' in result) {
@@ -198,14 +173,30 @@ export const useTVLData = () => {
       throw new Error('Unexpected result format')
     },
     enabled: !!actor,
-    staleTime: 10 * 60_000,
-    refetchInterval: 5 * 60_000,
+    staleTime: 10 * 60_000, // 10 minutes
+    refetchInterval: 5 * 60_000, // Refetch every 5 minutes
   })
 }
 
-export const useHoldings = () => {
-  const { actor } = useICPIBackend()
+export const useUserBalance = (actor: Actor | null, principal: string | null) => {
+  return useQuery({
+    queryKey: [QUERY_KEYS.USER_BALANCE, principal],
+    queryFn: async () => {
+      if (!actor || !principal) throw new Error('Actor or principal not available')
+      const icpiBalance = await actor.balance_of({ owner: principal, subaccount: [] })
+      // Also fetch ckUSDT balance if needed
+      return {
+        icpi: icpiBalance,
+        ckusdt: 0, // This would come from ckUSDT canister
+      }
+    },
+    enabled: !!actor && !!principal,
+    refetchInterval: 2 * 60_000, // Refetch every 2 minutes
+    staleTime: 60_000,
+  })
+}
 
+export const useHoldings = (actor: Actor | null) => {
   return useQuery({
     queryKey: [QUERY_KEYS.HOLDINGS],
     queryFn: async () => {
@@ -213,14 +204,12 @@ export const useHoldings = () => {
       return await actor.get_token_holdings()
     },
     enabled: !!actor,
-    refetchInterval: 2 * 60_000,
+    refetchInterval: 2 * 60_000, // Refetch every 2 minutes
     staleTime: 60_000,
   })
 }
 
-export const useAllocation = () => {
-  const { actor } = useICPIBackend()
-
+export const useAllocation = (actor: Actor | null) => {
   return useQuery({
     queryKey: [QUERY_KEYS.ALLOCATION],
     queryFn: async () => {
@@ -228,6 +217,7 @@ export const useAllocation = () => {
       const stateResult = await actor.get_index_state()
       const tvlResult = await actor.get_tvl_summary()
 
+      // Unwrap Result types
       if (!('Ok' in stateResult) || !('Ok' in tvlResult)) {
         throw new Error(
           'Err' in stateResult ? stateResult.Err :
@@ -236,321 +226,31 @@ export const useAllocation = () => {
         )
       }
 
+      // Calculate allocations based on state and TVL
       return calculateAllocations(stateResult.Ok, tvlResult.Ok)
     },
     enabled: !!actor,
-    refetchInterval: 2 * 60_000,
+    refetchInterval: 2 * 60_000, // Refetch every 2 minutes
     staleTime: 60_000,
   })
 }
 
-export const useTotalSupply = () => {
-  const { actor } = useICPIBackend()
-
-  return useQuery({
-    queryKey: [QUERY_KEYS.TOTAL_SUPPLY],
-    queryFn: async () => {
-      if (!actor) throw new Error('Actor not initialized')
-      const supply = await actor.icrc1_total_supply()
-      return Number(supply) / 100_000_000
-    },
-    enabled: !!actor,
-    staleTime: 10_000,
-    refetchInterval: 30_000,
-  })
-}
-
-export const useTokenMetadata = () => {
-  const { actor } = useICPIBackend()
-
-  return useQuery({
-    queryKey: [QUERY_KEYS.TOKEN_METADATA],
-    queryFn: async () => {
-      if (!actor) throw new Error('Actor not initialized')
-      const result = await actor.get_token_metadata()
-
-      if ('Ok' in result) {
-        return result.Ok
-      } else if ('Err' in result) {
-        throw new Error(result.Err)
-      }
-      throw new Error('Unexpected result format')
-    },
-    enabled: !!actor,
-    staleTime: Infinity,
-  })
-}
-
-export const useActualAllocations = () => {
-  const { actor } = useICPIBackend()
-  const agent = useAuthenticatedAgent()
-  const queryClient = useQueryClient()
-
-  return useQuery({
-    queryKey: [QUERY_KEYS.ACTUAL_ALLOCATIONS, ICPI_BACKEND_CANISTER_ID],
-    queryFn: async () => {
-      if (!actor || !agent) {
-        throw new Error('Actor or agent not initialized')
-      }
-
-      const [tokenMetadataResult, tvlDataResult] = await Promise.all([
-        actor.get_token_metadata(),
-        actor.get_tvl_summary(),
-      ])
-
-      if (!('Ok' in tokenMetadataResult) || !('Ok' in tvlDataResult)) {
-        throw new Error(
-          'Err' in tokenMetadataResult ? tokenMetadataResult.Err :
-          'Err' in tvlDataResult ? tvlDataResult.Err :
-          'Failed to fetch data'
-        )
-      }
-      const tokenMetadata = tokenMetadataResult.Ok
-      const tvlData = tvlDataResult.Ok
-
-      const balancePromises = tokenMetadata.map(async (token: any) => {
-        try {
-          const tokenActor = Actor.createActor(ICRC1_IDL, {
-            agent,
-            canisterId: token.canister_id.toString(),
-          })
-
-          const balance = await tokenActor.icrc1_balance_of({
-            owner: Principal.fromText(ICPI_BACKEND_CANISTER_ID),
-            subaccount: [],
-          })
-
-          return {
-            symbol: token.symbol,
-            balance: balance.toString(),
-            decimals: token.decimals,
-          }
-        } catch (error) {
-          console.error(`Error querying ${token.symbol} balance:`, error)
-          return {
-            symbol: token.symbol,
-            balance: '0',
-            decimals: token.decimals,
-          }
-        }
-      })
-
-      const balances = await Promise.all(balancePromises)
-
-      let indexState = queryClient.getQueryData([QUERY_KEYS.INDEX_STATE])
-
-      if (!indexState) {
-        const indexStateResult = await actor.get_index_state()
-        if (!('Ok' in indexStateResult)) {
-          throw new Error('Err' in indexStateResult ? indexStateResult.Err : 'Failed to get index state')
-        }
-        indexState = indexStateResult.Ok
-      }
-
-      const TRACKED_TOKENS = ['ALEX', 'ZERO', 'KONG', 'BOB']
-
-      const allocations = balances
-        .filter((bal) => TRACKED_TOKENS.includes(bal.symbol))
-        .map((bal) => {
-          const balance = BigInt(bal.balance)
-          const decimals = BigInt(10) ** BigInt(bal.decimals)
-          const balanceFloat = Number(balance) / Number(decimals)
-
-          const position = indexState.current_positions.find(
-            (p: any) => p.token[Object.keys(p.token)[0]] !== undefined &&
-                        Object.keys(p.token)[0] === bal.symbol
-          )
-
-          const usdValue = position ? position.usd_value : 0
-
-          const tvlEntry = tvlData.tokens.find(
-            (t: any) => Object.keys(t.token)[0] === bal.symbol
-          )
-          const targetPercent = tvlEntry ? tvlEntry.percentage : 0
-
-          return {
-            token: bal.symbol,
-            balance: balanceFloat,
-            value: usdValue,
-            decimals: bal.decimals,
-            targetPercent,
-          }
-        })
-
-      const ckusdtPosition = indexState.current_positions.find(
-        (p: any) => p.token.ckUSDT !== undefined
-      )
-
-      if (ckusdtPosition) {
-        allocations.push({
-          token: 'ckUSDT',
-          balance: Number(ckusdtPosition.balance) / 1_000_000,
-          value: ckusdtPosition.usd_value,
-          decimals: 6,
-          targetPercent: 0,
-        })
-      }
-
-      const totalValue = allocations.reduce((sum, a) => sum + a.value, 0)
-
-      const result = allocations.map((a) => ({
-        token: a.token,
-        value: a.value,
-        currentPercent: totalValue > 0 ? (a.value / totalValue) * 100 : 0,
-        targetPercent: a.targetPercent ?? 0,
-        deviation: totalValue > 0 ? (a.targetPercent ?? 0) - (a.value / totalValue) * 100 : 0,
-      }))
-
-      return result
-    },
-    enabled: !!actor && !!agent,
-    refetchInterval: 2 * 60_000,
-    staleTime: 60_000,
-  })
-}
-
-export const useUserWalletBalances = (userPrincipal: string) => {
-  const { actor } = useICPIBackend()
-  const agent = useAuthenticatedAgent()
-  const queryClient = useQueryClient()
-
-  return useQuery({
-    queryKey: [QUERY_KEYS.USER_WALLET_BALANCES, userPrincipal],
-    queryFn: async () => {
-      if (!actor || !userPrincipal || !agent) {
-        throw new Error('Actor, principal, or agent not initialized')
-      }
-
-      const tokenMetadataResult = await actor.get_token_metadata()
-
-      if (!('Ok' in tokenMetadataResult)) {
-        throw new Error('Err' in tokenMetadataResult ? tokenMetadataResult.Err : 'Failed to fetch token metadata')
-      }
-      const trackedTokensMetadata = tokenMetadataResult.Ok
-
-      const icpiMetadata = {
-        symbol: 'ICPI',
-        canister_id: Principal.fromText(ICPI_TOKEN_CANISTER_ID),
-        decimals: 8,
-      }
-
-      const ckusdtMetadata = {
-        symbol: 'ckUSDT',
-        canister_id: Principal.fromText(CKUSDT_CANISTER_ID),
-        decimals: 6,
-      }
-
-      const allTokensMetadata = [
-        icpiMetadata,
-        ckusdtMetadata,
-        ...trackedTokensMetadata,
-      ]
-
-      const balancePromises = allTokensMetadata.map(async (token: any) => {
-        try {
-          const tokenActor = Actor.createActor(ICRC1_IDL, {
-            agent,
-            canisterId: token.canister_id.toString(),
-          })
-
-          const balance = await tokenActor.icrc1_balance_of({
-            owner: Principal.fromText(userPrincipal),
-            subaccount: [],
-          })
-
-          const balanceStr = balance.toString()
-          const decimals = token.decimals
-          const balanceFormatted = Number(balanceStr) / Math.pow(10, decimals)
-
-          return {
-            symbol: token.symbol,
-            balance: balanceStr,
-            balanceFormatted,
-            decimals,
-            canisterId: token.canister_id.toString(),
-          }
-        } catch (error) {
-          console.error(`Error querying ${token.symbol} balance:`, error)
-          return {
-            symbol: token.symbol,
-            balance: '0',
-            balanceFormatted: 0,
-            decimals: token.decimals,
-            canisterId: token.canister_id.toString(),
-            error: error instanceof Error ? error.message : 'Query failed',
-          }
-        }
-      })
-
-      const balances = await Promise.all(balancePromises)
-
-      let enrichedBalances = balances
-      try {
-        let indexState = queryClient.getQueryData([QUERY_KEYS.INDEX_STATE])
-
-        if (!indexState) {
-          const indexStateResult = await actor.get_index_state()
-          if ('Ok' in indexStateResult) {
-            indexState = indexStateResult.Ok
-          }
-        }
-
-        if (indexState) {
-          enrichedBalances = balances.map(balance => {
-            const position = indexState.current_positions.find((p: any) => {
-              const tokenKey = Object.keys(p.token)[0]
-              return tokenKey === balance.symbol
-            })
-
-            if (position && balance.balanceFormatted > 0) {
-              const canisterBalance = Number(position.balance) / Math.pow(10, balance.decimals)
-              if (canisterBalance > 0) {
-                const pricePerToken = position.usd_value / canisterBalance
-                const usdValue = balance.balanceFormatted * pricePerToken
-                return { ...balance, usdValue }
-              }
-            }
-
-            return balance
-          })
-        }
-      } catch (error) {
-        console.error('Failed to fetch USD values:', error)
-      }
-
-      enrichedBalances.sort((a, b) => {
-        const aValue = a.usdValue || 0
-        const bValue = b.usdValue || 0
-        if (aValue !== bValue) return bValue - aValue
-        return b.balanceFormatted - a.balanceFormatted
-      })
-
-      return enrichedBalances as UserTokenBalance[]
-    },
-    enabled: !!actor && !!userPrincipal && !!agent,
-    refetchInterval: 30_000,
-    staleTime: 15_000,
-    retry: 2,
-  })
-}
-
-// ===== MUTATION HOOKS =====
-
-export const useMintICPI = () => {
-  const { actor } = useICPIBackend()
-  const agent = useAuthenticatedAgent()
+// Mutation hooks
+export const useMintICPI = (actor: Actor | null, agent: HttpAgent | null) => {
   const queryClient = useQueryClient()
 
   return useMutation({
     mutationFn: async (ckusdtAmount: number) => {
       if (!actor || !agent) throw new Error('Actor or agent not initialized')
 
-      const amountRaw = BigInt(Math.floor(ckusdtAmount * 1e6))
-      const feeAmount = BigInt(100_000)
-      const ckusdtTransferFee = BigInt(10_000)
+      const amountRaw = BigInt(Math.floor(ckusdtAmount * 1e6)) // ckUSDT has 6 decimals
+      const feeAmount = BigInt(100_000) // 0.1 ckUSDT backend fee
+      const ckusdtTransferFee = BigInt(10_000) // 0.01 ckUSDT ICRC1 transfer fee
 
+      // Total approval: mint amount + backend fee + buffer for 2 transfers (fee collection + deposit)
       const totalApproval = amountRaw + feeAmount + (ckusdtTransferFee * BigInt(2))
 
+      // Step 0: Approve ICPI backend to spend ckUSDT
       const ckusdtActor = Actor.createActor(ICRC1_IDL, {
         agent,
         canisterId: CKUSDT_CANISTER_ID,
@@ -578,6 +278,7 @@ export const useMintICPI = () => {
         throw new Error(`Approval failed: ${JSON.stringify(approveResult.Err)}`)
       }
 
+      // Phase 1: Initiate mint (returns mint_id)
       const initResult = await actor.initiate_mint(amountRaw)
 
       if ('Err' in initResult) {
@@ -585,6 +286,7 @@ export const useMintICPI = () => {
       }
       const mintId = initResult.Ok
 
+      // Phase 2: Complete mint (returns ICPI amount)
       const completeResult = await actor.complete_mint(mintId)
 
       if ('Err' in completeResult) {
@@ -594,6 +296,7 @@ export const useMintICPI = () => {
       return completeResult.Ok
     },
     onSuccess: () => {
+      // Invalidate relevant queries
       queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.INDEX_STATE] })
       queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.USER_BALANCE] })
       queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.HOLDINGS] })
@@ -602,20 +305,19 @@ export const useMintICPI = () => {
   })
 }
 
-export const useRedeemICPI = () => {
-  const { actor } = useICPIBackend()
-  const agent = useAuthenticatedAgent()
+export const useRedeemICPI = (actor: Actor | null, agent: HttpAgent | null) => {
   const queryClient = useQueryClient()
 
   return useMutation({
     mutationFn: async (icpiAmount: number) => {
       if (!actor || !agent) throw new Error('Actor or agent not initialized')
 
-      const amountRaw = BigInt(Math.floor(icpiAmount * 1e8))
-      const feeAmount = BigInt(100_000)
-      const ckusdtTransferFee = BigInt(10_000)
-      const icpiTransferFee = BigInt(10_000)
+      const amountRaw = BigInt(Math.floor(icpiAmount * 1e8)) // ICPI has 8 decimals
+      const feeAmount = BigInt(100_000) // 0.1 ckUSDT backend fee (6 decimals)
+      const ckusdtTransferFee = BigInt(10_000) // 0.01 ckUSDT ICRC1 transfer fee
+      const icpiTransferFee = BigInt(10_000) // ICPI transfer fee (8 decimals, but numerically 10000)
 
+      // Step 1: Approve backend to collect fee (ckUSDT)
       const ckusdtActor = Actor.createActor(ICRC1_IDL, {
         agent,
         canisterId: CKUSDT_CANISTER_ID,
@@ -629,7 +331,7 @@ export const useRedeemICPI = () => {
           owner: icpiBackendPrincipal,
           subaccount: [],
         },
-        amount: feeAmount + ckusdtTransferFee,
+        amount: feeAmount + ckusdtTransferFee, // Fee + transfer fee
         expected_allowance: [],
         expires_at: [],
         fee: [ckusdtTransferFee],
@@ -643,6 +345,7 @@ export const useRedeemICPI = () => {
         throw new Error(`Fee approval failed: ${JSON.stringify(approveResult.Err)}`)
       }
 
+      // Step 2: Transfer ICPI to backend (this burns it since backend is minting account)
       const icpiActor = Actor.createActor(ICRC1_IDL, {
         agent,
         canisterId: ICPI_TOKEN_CANISTER_ID,
@@ -666,6 +369,7 @@ export const useRedeemICPI = () => {
         throw new Error(`ICPI transfer failed: ${JSON.stringify(transferResult.Err)}`)
       }
 
+      // Step 3: Call atomic burn_icpi (backend will verify burn and send redemption tokens)
       const burnResult = await actor.burn_icpi(amountRaw)
 
       if ('Err' in burnResult) {
@@ -683,8 +387,7 @@ export const useRedeemICPI = () => {
   })
 }
 
-export const useManualRebalance = () => {
-  const { actor } = useICPIBackend()
+export const useManualRebalance = (actor: Actor | null) => {
   const queryClient = useQueryClient()
 
   return useMutation({
@@ -701,8 +404,8 @@ export const useManualRebalance = () => {
   })
 }
 
-export const useTransferToken = () => {
-  const agent = useAuthenticatedAgent()
+// Hook for transferring ICRC1 tokens
+export const useTransferToken = (agent: HttpAgent | null) => {
   const queryClient = useQueryClient()
 
   return useMutation({
@@ -719,35 +422,43 @@ export const useTransferToken = () => {
     }) => {
       if (!agent) throw new Error('Agent not initialized')
 
+      // STEP 1: Create actor for the token canister
       const tokenActor = Actor.createActor(ICRC1_IDL, {
         agent,
         canisterId: tokenCanisterId,
       })
 
+      // STEP 2: Query fee from token canister
       const fee = await tokenActor.icrc1_fee()
 
+      // STEP 3: Convert amount to raw units using string manipulation
+      // to avoid floating-point precision loss
       const [wholePart, decimalPart = ''] = amount.split('.')
       const paddedDecimal = decimalPart.padEnd(decimals, '0').slice(0, decimals)
       const amountStr = wholePart + paddedDecimal
       const amountRaw = BigInt(amountStr)
 
+      // STEP 4: Prepare transfer args (ICRC1 standard)
       const transferArgs = {
         to: {
           owner: Principal.fromText(recipient),
           subaccount: [],
         },
         amount: amountRaw,
-        fee: [fee],
-        memo: [],
-        from_subaccount: [],
-        created_at_time: [],
+        fee: [fee], // Always include fee
+        memo: [], // Optional
+        from_subaccount: [], // Optional
+        created_at_time: [], // Optional
       }
 
+      // STEP 5: Call icrc1_transfer
       const result = await tokenActor.icrc1_transfer(transferArgs)
 
+      // STEP 6: Handle Result type (Ok/Err)
       if ('Ok' in result) {
-        return result.Ok
+        return result.Ok // Block index
       } else if ('Err' in result) {
+        // Parse error for user-friendly message
         const errMsg = JSON.stringify(result.Err)
         if (errMsg.includes('InsufficientFunds')) {
           throw new Error('Insufficient funds (remember to account for transfer fee)')
@@ -759,21 +470,360 @@ export const useTransferToken = () => {
       throw new Error('Unexpected result format')
     },
     onSuccess: () => {
+      // Invalidate user wallet balances to trigger refresh
       queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.USER_WALLET_BALANCES] })
     },
   })
 }
 
-// Helper function
+// Get token metadata from backend
+export const useTokenMetadata = (actor: Actor | null) => {
+  return useQuery({
+    queryKey: [QUERY_KEYS.TOKEN_METADATA],
+    queryFn: async () => {
+      if (!actor) throw new Error('Actor not initialized')
+      const result = await actor.get_token_metadata()
+
+      // Handle Result type - unwrap Ok/Err variant
+      if ('Ok' in result) {
+        return result.Ok
+      } else if ('Err' in result) {
+        throw new Error(result.Err)
+      }
+      throw new Error('Unexpected result format')
+    },
+    enabled: !!actor,
+    staleTime: Infinity, // Token metadata never changes
+  })
+}
+
+// Get ICPI total supply
+export const useTotalSupply = (actor: Actor | null) => {
+  return useQuery({
+    queryKey: [QUERY_KEYS.TOTAL_SUPPLY],
+    queryFn: async () => {
+      if (!actor) throw new Error('Actor not initialized')
+      const supply = await actor.icrc1_total_supply()
+      // Convert from e8 to human-readable (8 decimals)
+      return Number(supply) / 100_000_000
+    },
+    enabled: !!actor,
+    staleTime: 10_000, // 10 seconds
+    refetchInterval: 30_000, // Refetch every 30s
+  })
+}
+
+// Query token balances directly from token canisters
+export const useActualAllocations = (
+  icpiActor: Actor | null,
+  icpiCanisterId: string | null,
+  agent: HttpAgent | null
+) => {
+  const queryClient = useQueryClient()
+  return useQuery({
+    queryKey: [QUERY_KEYS.ACTUAL_ALLOCATIONS, icpiCanisterId],
+    queryFn: async () => {
+      if (!icpiActor || !icpiActor || !icpiCanisterId || !agent) {
+        throw new Error('Actor, canister ID, or agent not initialized')
+      }
+
+      // Get token metadata and TVL data
+      const [tokenMetadataResult, tvlDataResult] = await Promise.all([
+        icpiActor.get_token_metadata(),
+        icpiActor.get_tvl_summary(),
+      ])
+
+      // Unwrap Result types
+      if (!('Ok' in tokenMetadataResult) || !('Ok' in tvlDataResult)) {
+        throw new Error(
+          'Err' in tokenMetadataResult ? tokenMetadataResult.Err :
+          'Err' in tvlDataResult ? tvlDataResult.Err :
+          'Failed to fetch data'
+        )
+      }
+      const tokenMetadata = tokenMetadataResult.Ok
+      const tvlData = tvlDataResult.Ok
+
+      // Query balances directly from token canisters (true query calls!)
+      const balancePromises = tokenMetadata.map(async (token: any) => {
+        try {
+          const tokenActor = Actor.createActor(ICRC1_IDL, {
+            agent,
+            canisterId: token.canister_id.toString(),
+          })
+
+          const balance = await tokenActor.icrc1_balance_of({
+            owner: Principal.fromText(icpiCanisterId),
+            subaccount: [],
+          })
+
+          return {
+            symbol: token.symbol,
+            balance: balance.toString(),
+            decimals: token.decimals,
+          }
+        } catch (error) {
+          console.error(`Error querying ${token.symbol} balance:`, error)
+          return {
+            symbol: token.symbol,
+            balance: '0',
+            decimals: token.decimals,
+          }
+        }
+      })
+
+      const balances = await Promise.all(balancePromises)
+
+      // Calculate USD values and percentages
+      // Get current positions from backend for prices
+      let indexState = queryClient.getQueryData([QUERY_KEYS.INDEX_STATE])
+
+      if (!indexState) {
+        // Fallback: call it if somehow not cached yet
+        const indexStateResult = await icpiActor.get_index_state()
+        if (!('Ok' in indexStateResult)) {
+          throw new Error('Err' in indexStateResult ? indexStateResult.Err : 'Failed to get index state')
+        }
+        indexState = indexStateResult.Ok
+      }
+
+      // Track index tokens + ckUSDT (cash position for rebalancing)
+      const TRACKED_TOKENS = ['ALEX', 'ZERO', 'KONG', 'BOB']
+
+      const allocations = balances
+        .filter((bal) => TRACKED_TOKENS.includes(bal.symbol))
+        .map((bal) => {
+          const balance = BigInt(bal.balance)
+          const decimals = BigInt(10) ** BigInt(bal.decimals)
+          const balanceFloat = Number(balance) / Number(decimals)
+
+          // Find matching position for price
+          const position = indexState.current_positions.find(
+            (p: any) => p.token[Object.keys(p.token)[0]] !== undefined &&
+                        Object.keys(p.token)[0] === bal.symbol
+          )
+
+          const usdValue = position ? position.usd_value : 0
+
+          // Find target from TVL data
+          const tvlEntry = tvlData.tokens.find(
+            (t: any) => Object.keys(t.token)[0] === bal.symbol
+          )
+          const targetPercent = tvlEntry ? tvlEntry.percentage : 0
+
+          return {
+            token: bal.symbol,
+            balance: balanceFloat,
+            value: usdValue,
+            decimals: bal.decimals,
+            targetPercent,
+          }
+        })
+
+      // Add ckUSDT position (cash held for rebalancing)
+      const ckusdtPosition = indexState.current_positions.find(
+        (p: any) => p.token.ckUSDT !== undefined
+      )
+
+      if (ckusdtPosition) {
+        allocations.push({
+          token: 'ckUSDT',
+          balance: Number(ckusdtPosition.balance) / 1_000_000, // 6 decimals
+          value: ckusdtPosition.usd_value,
+          decimals: 6,
+          targetPercent: 0, // ckUSDT is cash, not a target allocation
+        })
+      }
+
+      // Calculate actual percentages (including ckUSDT to show total portfolio composition)
+      const totalValue = allocations.reduce((sum, a) => sum + a.value, 0)
+
+      const result = allocations.map((a) => ({
+        token: a.token,
+        value: a.value,
+        currentPercent: totalValue > 0 ? (a.value / totalValue) * 100 : 0,
+        targetPercent: a.targetPercent ?? 0,
+        deviation: totalValue > 0 ? (a.targetPercent ?? 0) - (a.value / totalValue) * 100 : 0,
+      }))
+
+      return result
+    },
+    enabled: !!icpiActor && !!icpiCanisterId && !!agent,
+    refetchInterval: 2 * 60_000, // Refetch every 2 minutes (query calls are fast but no need to spam)
+    staleTime: 60_000,
+  })
+}
+
+// Hook to get tracked tokens
+export const useTrackedTokens = (actor: Actor | null) => {
+  return useQuery({
+    queryKey: [QUERY_KEYS.TRACKED_TOKENS],
+    queryFn: async () => {
+      if (!actor) return ['ALEX', 'ZERO', 'KONG', 'BOB'] // Fallback
+      return await actor.get_tracked_tokens()
+    },
+    staleTime: 5 * 60_000, // Cache for 5 minutes
+  })
+}
+
+// Hook to get user wallet balances for all tokens
+export const useUserWalletBalances = (
+  actor: Actor | null,
+  userPrincipal: string | null,
+  agent: HttpAgent | null
+) => {
+  const queryClient = useQueryClient()
+  return useQuery({
+    queryKey: [QUERY_KEYS.USER_WALLET_BALANCES, userPrincipal],
+    queryFn: async () => {
+      if (!actor || !userPrincipal || !agent) {
+        throw new Error('Actor, principal, or agent not initialized')
+      }
+
+      // STEP 1: Get token metadata for all tracked tokens
+      const tokenMetadataResult = await actor.get_token_metadata()
+
+      if (!('Ok' in tokenMetadataResult)) {
+        throw new Error('Err' in tokenMetadataResult ? tokenMetadataResult.Err : 'Failed to fetch token metadata')
+      }
+      const trackedTokensMetadata = tokenMetadataResult.Ok
+
+      // STEP 2: Add ICPI token metadata (the index token itself)
+      const icpiMetadata = {
+        symbol: 'ICPI',
+        canister_id: Principal.fromText(ICPI_TOKEN_CANISTER_ID),
+        decimals: 8, // ICPI has 8 decimals
+      }
+
+      // STEP 3: Add ckUSDT metadata (special case - rebalancing currency)
+      const ckusdtMetadata = {
+        symbol: 'ckUSDT',
+        canister_id: Principal.fromText(CKUSDT_CANISTER_ID),
+        decimals: 6, // ckUSDT has 6 decimals
+      }
+
+      // STEP 4: Combine all token metadata
+      const allTokensMetadata = [
+        icpiMetadata,
+        ckusdtMetadata,
+        ...trackedTokensMetadata,
+      ]
+
+      // STEP 5: Query balances for all tokens in parallel
+      const balancePromises = allTokensMetadata.map(async (token: any) => {
+        try {
+          // Create ICRC1 actor for each token canister
+          const tokenActor = Actor.createActor(ICRC1_IDL, {
+            agent,
+            canisterId: token.canister_id.toString(),
+          })
+
+          // Query balance using ICRC1 standard
+          const balance = await tokenActor.icrc1_balance_of({
+            owner: Principal.fromText(userPrincipal),
+            subaccount: [],
+          })
+
+          // Convert to human-readable format
+          const balanceStr = balance.toString()
+          const decimals = token.decimals
+          const balanceFormatted = Number(balanceStr) / Math.pow(10, decimals)
+
+          return {
+            symbol: token.symbol,
+            balance: balanceStr,
+            balanceFormatted,
+            decimals,
+            canisterId: token.canister_id.toString(),
+          }
+        } catch (error) {
+          console.error(`Error querying ${token.symbol} balance:`, error)
+          // Return zero balance with error flag so UI can show warning
+          return {
+            symbol: token.symbol,
+            balance: '0',
+            balanceFormatted: 0,
+            decimals: token.decimals,
+            canisterId: token.canister_id.toString(),
+            error: error instanceof Error ? error.message : 'Query failed',
+          }
+        }
+      })
+
+      // STEP 6: Await all balance queries
+      const balances = await Promise.all(balancePromises)
+
+      // STEP 7: Get USD values from index state
+      let enrichedBalances = balances
+      try {
+        // Try cache first
+        let indexState = queryClient.getQueryData([QUERY_KEYS.INDEX_STATE])
+
+        if (!indexState) {
+          // Fallback only if cache miss
+          const indexStateResult = await actor.get_index_state()
+          if ('Ok' in indexStateResult) {
+            indexState = indexStateResult.Ok
+          }
+        }
+
+        if (indexState) {
+
+          // Enrich balances with USD values
+          enrichedBalances = balances.map(balance => {
+            // Find matching position for USD value
+            const position = indexState.current_positions.find((p: any) => {
+              const tokenKey = Object.keys(p.token)[0]
+              return tokenKey === balance.symbol
+            })
+
+            if (position && balance.balanceFormatted > 0) {
+              // Calculate USD value: (user_balance / canister_balance) * canister_usd_value
+              const canisterBalance = Number(position.balance) / Math.pow(10, balance.decimals)
+              if (canisterBalance > 0) {
+                const pricePerToken = position.usd_value / canisterBalance
+                const usdValue = balance.balanceFormatted * pricePerToken
+                return { ...balance, usdValue }
+              }
+            }
+
+            return balance
+          })
+        }
+      } catch (error) {
+        console.error('Failed to fetch USD values:', error)
+        // Continue without USD values if this fails
+      }
+
+      // STEP 8: Sort by USD value (highest first), then by balance
+      enrichedBalances.sort((a, b) => {
+        const aValue = a.usdValue || 0
+        const bValue = b.usdValue || 0
+        if (aValue !== bValue) return bValue - aValue
+        return b.balanceFormatted - a.balanceFormatted
+      })
+
+      return enrichedBalances as UserTokenBalance[]
+    },
+    enabled: !!actor && !!userPrincipal && !!agent,
+    refetchInterval: 30_000, // Refetch every 30 seconds (query calls are fast)
+    staleTime: 15_000,       // Consider stale after 15 seconds
+    retry: 2,
+  })
+}
+
+// Helper function to calculate allocations
 function calculateAllocations(state: any, tvl: any) {
+  // This would process the raw data into the format needed by the UI
   const tokens = ['ALEX', 'ZERO', 'KONG', 'BOB']
   return tokens.map(token => {
+    // Calculate based on actual holdings vs TVL targets
     return {
       token,
-      value: 0,
-      currentPercent: 0,
-      targetPercent: 0,
-      deviation: 0,
+      value: 0, // Calculate from state
+      currentPercent: 0, // Calculate
+      targetPercent: 0, // From TVL data
+      deviation: 0, // Current - Target
     }
   })
 }
